Dactyl keymap - remove tap mods on the Sym layer.
Mod-taps do not work with shifted keys, so @() and :" did not work
on the Sym layer.

The solution should be to use tap dance to emulate mod tap, but it
also requires switching the layout from json file to keymap.c, it
will be less convenient to edit it.

I think mods on the Sym layer are not very critical, so I'll try
to use the layout without them.